+++ addons/fray/src/collision/2d/hitbox_2d.gd
@@ -5,7 +5,7 @@ extends Area2D
 ## 2D area intended to detect combat interactions.
 ##
 ## The hitbox node doesn't provide much functionality out of the box.
-## It serves as a template you can expand upon through the use of [FrayHitboxAttributes]
+## It serves as a template you can expand upon through the use of [FrayHitboxAtrribute]
 
 ## Emitted when the received [kbd]hitbox[/kbd] enters this hitbox. Requires monitoring to be set to [code]true[/code].
 signal hitbox_entered(hitbox: FrayHitbox2D)
@@ -16,12 +16,12 @@ signal hitbox_exited(hitbox: FrayHitbox2D)
 ## If true then hitboxes that share the same source as this one will still be detected
 @export var detect_source_hitboxes: bool = false
 
-## The [FrayHitboxAttributes] assigned to this hitbox
-@export var attributes: FrayHitboxAttributes = null:
+## The [FrayHitboxAtrribute] assigned to this hitbox
+@export var atrribute: FrayHitboxAtrribute = null:
 	set(value):
-		attributes = value
+		atrribute = value
 		
-		if attributes != null:
+		if atrribute != null:
 			_update_collision_colors()
 		
 		update_configuration_warnings()
@@ -49,8 +49,8 @@ func _ready() -> void:
 
 func _get_configuration_warnings() -> PackedStringArray:
 	var warnings: PackedStringArray = []
-	if attributes == null:
-		warnings.append("Hitboxes without attributes are effectively just Area2Ds. Consider giving this node a FrayHitboxAttributes resource.")
+	if atrribute == null:
+		warnings.append("Hitboxes without atrribute are effectively just Area2Ds. Consider giving this node a FrayHitboxAtrribute resource.")
 	return warnings
 
 ## Returns a list of intersecting [FrayHitbox2D]s.
@@ -102,16 +102,16 @@ func can_detect(hitbox: FrayHitbox2D) -> bool:
 		not _hitbox_exceptions.has(hitbox)
 		and not _source_exceptions.has(hitbox.source)
 		and detect_source_hitboxes or hitbox.source != source
-		and attributes.allows_detection_of(hitbox.attributes)
-			if attributes != null else true
+		and atrribute.allows_detection_of(hitbox.atrribute)
+			if atrribute != null else true
 		)
 	
 
 func _update_collision_colors() -> void:
-	if attributes != null:
+	if atrribute != null:
 		for node in get_children():
 			if node is CollisionShape2D:
-				node.debug_color = attributes.get_color()
+				node.debug_color = atrribute.get_color()
 	
 	
 func _on_area_entered(area: Area2D) -> void: